# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BookVerse Checkout Service API
  description: |
    ## Overview
    
    The BookVerse Checkout Service handles the complete order processing workflow including order creation, payment processing, inventory coordination, and order fulfillment. Built with enterprise-grade reliability and security features.
    
    ## Key Features
    
    - **Complete Order Management**: End-to-end order lifecycle from creation to fulfillment
    - **Multi-Gateway Payment Processing**: Support for multiple payment providers with failover
    - **Inventory Integration**: Real-time inventory checking and reservation
    - **Idempotency Support**: Safe retry mechanisms for critical operations
    - **Fraud Prevention**: Advanced fraud detection and risk assessment
    - **Audit Trail**: Comprehensive order and payment audit logging
    
    ## Payment Support
    
    - **Credit/Debit Cards**: Visa, Mastercard, American Express, Discover
    - **Digital Wallets**: PayPal, Apple Pay, Google Pay, Amazon Pay
    - **Bank Transfers**: ACH, wire transfers, SEPA (EU)
    - **Buy Now Pay Later**: Affirm, Klarna, Afterpay integration
    - **Cryptocurrency**: Bitcoin, Ethereum (via partner providers)
    
    ## Security & Compliance
    
    - **PCI DSS Level 1** compliant payment processing
    - **SOX compliance** for financial audit trails
    - **GDPR/CCPA** compliant data handling
    - **3D Secure 2.0** for enhanced card authentication
...
    
    ## Order States
    
...
servers:
...
security:
  - BearerAuth: []
paths:
  /orders:
    get:
      summary: List orders with filtering and pagination
      description: |
        Retrieve a paginated list of orders with comprehensive filtering options for order management and customer service.
        
        **Filtering Capabilities:**
        - Order status and date ranges
        - Customer information and order values
        - Payment status and methods
        - Geographic and shipping information
        
        **Use Cases:**
        - Customer order history display
        - Administrative order management
        - Financial reconciliation and reporting
        - Customer service inquiry handling
        
        **Performance:**
        - Optimized queries with database indexing
        - Response caching for frequently accessed data
        - Pagination for large result sets
      operationId: listOrders
      tags:
        - Orders
      parameters:
        - name: customer_id
          in: query
          description: Filter orders by customer ID
          schema:
            type: string
            maxLength: 100
            example: "customer_12345"
        
        - name: status
          in: query
          description: Filter orders by status
          schema:
            type: string
            enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned]
            example: confirmed
        
        - name: payment_status
          in: query
          description: Filter orders by payment status
          schema:
            type: string
            enum: [pending, authorized, captured, failed, refunded, partially_refunded]
            example: captured
        
        - name: date_from
          in: query
          description: Filter orders from this date (inclusive)
          schema:
            type: string
            format: date
            example: "2024-01-01"
        
        - name: date_to
          in: query
          description: Filter orders to this date (inclusive)
          schema:
            type: string
            format: date
            example: "2024-01-31"
        
        - name: min_amount
          in: query
          description: Minimum order amount filter
          schema:
            type: number
            format: decimal
            minimum: 0
            example: 25.00
        
        - name: max_amount
          in: query
          description: Maximum order amount filter
          schema:
            type: number
            format: decimal
            minimum: 0
            example: 500.00
        
        - name: payment_method
          in: query
          description: Filter by payment method
          schema:
            type: string
            enum: [credit_card, debit_card, paypal, apple_pay, google_pay, bank_transfer, crypto]
            example: credit_card
        
        - name: sort_by
          in: query
          description: Field to sort results by
          schema:
            type: string
            enum: [created_at, updated_at, total_amount, status]
            default: created_at
            example: created_at
        
        - name: sort_order
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
            example: desc
        
        - name: limit
          in: query
          description: Maximum number of results per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        
        - name: include_items
          in: query
          description: Include order items in response
          schema:
            type: boolean
            default: false
            example: true

      responses:
        '200':
          description: Orders retrieved successfully
          headers:
            X-Total-Count:
              description: Total number of orders matching criteria
              schema:
                type: integer
                example: 2847
            X-Page-Count:
              description: Total number of pages
              schema:
                type: integer
                example: 143
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
                  summary:
                    $ref: '#/components/schemas/OrderSummary'
              examples:
                order_list:
                  summary: List of customer orders
                  value:
                    data:
                      - id: "order_789abc123def"
                        customer_id: "customer_12345"
                        status: "delivered"
                        payment_status: "captured"
                        total_amount: 89.97
                        currency: "USD"
                        created_at: "2024-01-20T10:30:00Z"
                        updated_at: "2024-01-25T14:22:00Z"
                        item_count: 3
                        shipping_address:
                          name: "John Doe"
                          street: "123 Main St"
                          city: "New York"
                          state: "NY"
                          postal_code: "10001"
                          country: "US"
                    pagination:
                      limit: 20
                      offset: 0
                      total: 2847
                      has_more: true
                    summary:
                      total_orders: 2847
                      total_value: 156742.33
                      avg_order_value: 55.12

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new order
      description: |
        Create a new order with comprehensive validation, inventory checking, and idempotency support for safe retries.
        
        **Order Creation Process:**
        1. **Input Validation**: Validate customer data and order items
        2. **Inventory Check**: Verify book availability and reserve stock
        3. **Pricing Calculation**: Calculate totals, taxes, and shipping
        4. **Risk Assessment**: Fraud detection and risk scoring
        5. **Order Creation**: Create order with unique identifier
        6. **Audit Logging**: Record order creation for compliance
        
        **Business Rules:**
        - Minimum order value: $5.00
        - Maximum order value: $10,000.00
        - Maximum 50 items per order
        - Valid shipping address required
        - Customer information must be complete
        
        **Idempotency:**
        - Use `Idempotency-Key` header for safe retries
        - Duplicate requests return the same order
        - Keys expire after 24 hours
      operationId: createOrder
      tags:
        - Orders
      parameters:
        - name: Idempotency-Key
          in: header
          description: |
            Idempotency key for safe order creation retries.
            Use a unique UUID for each order attempt.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              simple_order:
                summary: Simple single-book order
                value:
                  customer_id: "customer_12345"
                  items:
                    - book_id: 123
                      quantity: 1
                  shipping_address:
                    name: "John Doe"
                    email: "john@example.com"
                    phone: "+1-555-123-4567"
                    street: "123 Main St"
                    city: "New York"
                    state: "NY"
                    postal_code: "10001"
                    country: "US"
                  billing_address:
                    name: "John Doe"
                    street: "123 Main St"
                    city: "New York"
                    state: "NY"
                    postal_code: "10001"
                    country: "US"
                  notes: "Please handle with care"
              
              multi_item_order:
                summary: Multi-item order with discount
                value:
                  customer_id: "customer_67890"
                  items:
                    - book_id: 123
                      quantity: 2
                    - book_id: 456
                      quantity: 1
                    - book_id: 789
                      quantity: 3
                  shipping_address:
                    name: "Jane Smith"
                    email: "jane@example.com"
                    phone: "+1-555-987-6543"
                    street: "456 Oak Avenue"
                    city: "Los Angeles"
                    state: "CA"
                    postal_code: "90210"
                    country: "US"
                  billing_address:
                    name: "Jane Smith"
                    street: "456 Oak Avenue"
                    city: "Los Angeles"
                    state: "CA"
                    postal_code: "90210"
                    country: "US"
                  discount_code: "SAVE20"
                  notes: "Gift wrap requested"

      responses:
        '201':
          description: Order created successfully
          headers:
            Location:
              description: URL of the created order
              schema:
                type: string
                format: uri
                example: "/orders/order_789abc123def"
            Idempotency-Key:
              description: Echo of the idempotency key used
              schema:
                type: string
                example: "550e8400-e29b-41d4-a716-446655440000"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
                  processing_info:
                    type: object
                    properties:
                      inventory_reserved:
                        type: boolean
                        description: Whether inventory was successfully reserved
                        example: true
                      risk_score:
                        type: number
                        format: float
                        description: Order risk assessment score (0-1)
                        example: 0.15
                      estimated_shipping_date:
                        type: string
                        format: date
                        description: Estimated shipping date
                        example: "2024-01-27"
                      processing_time_ms:
                        type: integer
                        description: Order processing time
                        example: 234
              examples:
                created_order:
                  summary: Successfully created order
                  value:
                    data:
                      id: "order_789abc123def"
                      customer_id: "customer_12345"
                      status: "pending"
                      payment_status: "pending"
                      items:
                        - book_id: 123
                          title: "Clean Code"
                          author: "Robert Martin"
                          quantity: 1
                          unit_price: 35.99
                          total_price: 35.99
                      subtotal: 35.99
                      tax_amount: 3.24
                      shipping_amount: 7.99
                      discount_amount: 0.00
                      total_amount: 47.22
                      currency: "USD"
                      created_at: "2024-01-25T16:30:00Z"
                      estimated_delivery: "2024-02-01T00:00:00Z"
                    processing_info:
                      inventory_reserved: true
                      risk_score: 0.15
                      estimated_shipping_date: "2024-01-27"
                      processing_time_ms: 234

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Inventory conflict or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                inventory_conflict:
                  summary: Insufficient inventory
                  value:
                    type: "https://docs.bookverse.com/errors/insufficient-inventory"
                    title: "Insufficient Inventory"
                    status: 409
                    detail: "Book 'Clean Code' has only 2 copies available, but 5 were requested"
                    instance: "/orders"
                    inventory_details:
                      book_id: 123
                      available: 2
                      requested: 5
                
                idempotency_conflict:
                  summary: Idempotency key reused with different data
                  value:
                    type: "https://docs.bookverse.com/errors/idempotency-conflict"
                    title: "Idempotency Conflict"
                    status: 409
                    detail: "Idempotency key already used with different order data"
                    instance: "/orders"
                    existing_order_id: "order_456def789abc"

        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{order_id}:
    parameters:
      - name: order_id
        in: path
        required: true
        description: Unique order identifier
        schema:
          type: string
          pattern: '^order_[a-zA-Z0-9]{12}$'
          example: "order_789abc123def"

    get:
      summary: Get order details
      description: |
        Retrieve comprehensive order information including items, payment status, shipping details, and order history.
        
        **Information Included:**
        - Complete order details and current status
        - All order items with pricing breakdown
        - Payment information and transaction history
        - Shipping address and tracking information
        - Order timeline and status changes
        - Customer information (if authorized)
        
        **Authorization:**
        - Customers can view their own orders
        - Admin users can view any order
        - Customer service can view with proper permissions
        
        **Performance:**
        - Cached responses for completed orders
        - Real-time data for active orders
        - Optimized queries for order details
      operationId: getOrder
      tags:
        - Orders
      parameters:
        - name: include_history
          in: query
          description: Include order status change history
          schema:
            type: boolean
            default: false
            example: true
        
        - name: include_payments
          in: query
          description: Include payment transaction details
          schema:
            type: boolean
            default: false
            example: true

      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/OrderDetailed'
                  payment_history:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaymentTransaction'
                    description: Included only when include_payments=true
                  status_history:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderStatusChange'
                    description: Included only when include_history=true
              examples:
                detailed_order:
                  summary: Complete order with all details
                  value:
                    data:
                      id: "order_789abc123def"
                      customer_id: "customer_12345"
                      status: "processing"
                      payment_status: "captured"
                      items:
                        - book_id: 123
                          title: "Clean Code: A Handbook of Agile Software Craftsmanship"
                          author: "Robert C. Martin"
                          isbn: "978-0132350884"
                          quantity: 2
                          unit_price: 35.99
                          total_price: 71.98
                        - book_id: 456
                          title: "Design Patterns"
                          author: "Gang of Four"
                          isbn: "978-0201633612"
                          quantity: 1
                          unit_price: 54.99
                          total_price: 54.99
                      subtotal: 126.97
                      tax_amount: 11.43
                      shipping_amount: 9.99
                      discount_amount: 12.70
                      total_amount: 135.69
                      currency: "USD"
                      shipping_address:
                        name: "John Doe"
                        email: "john@example.com"
                        phone: "+1-555-123-4567"
                        street: "123 Main St"
                        city: "New York"
                        state: "NY"
                        postal_code: "10001"
                        country: "US"
                      tracking_number: "1Z999AA1234567890"
                      estimated_delivery: "2024-02-01T00:00:00Z"
                      created_at: "2024-01-25T16:30:00Z"
                      updated_at: "2024-01-26T09:15:00Z"

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Access denied - Cannot view this order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://docs.bookverse.com/errors/order-access-denied"
                title: "Order Access Denied"
                status: 403
                detail: "You do not have permission to view this order"
                instance: "/orders/order_789abc123def"
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://docs.bookverse.com/errors/order-not-found"
                title: "Order Not Found"
                status: 404
                detail: "No order found with ID order_789abc123def"
                instance: "/orders/order_789abc123def"
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Update order status or details
      description: |
        Update order information including status changes, shipping details, and customer information with comprehensive validation and audit logging.
        
        **Updatable Fields:**
        - Order status (with business rule validation)
        - Shipping address (before processing)
        - Customer notes and special instructions
        - Internal notes (admin only)
        - Delivery preferences
        
        **Business Rules:**
        - Status transitions must follow valid state machine
        - Shipping address cannot be changed after processing
        - Payment information cannot be modified directly
        - Cancellation requires authorization and inventory handling
        
        **Audit Trail:**
        - All changes are logged with timestamps
        - User attribution for all modifications
        - Compliance tracking for financial changes
      operationId: updateOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
            examples:
              status_update:
                summary: Update order status
                value:
                  status: "shipped"
                  tracking_number: "1Z999AA1234567890"
                  notes: "Package shipped via UPS Ground"
              
              address_update:
                summary: Update shipping address
                value:
                  shipping_address:
                    name: "John Doe"
                    street: "456 New Address St"
                    city: "Boston"
                    state: "MA"
                    postal_code: "02101"
                    country: "US"
                  notes: "Address updated per customer request"
              
              cancellation:
                summary: Cancel order
                value:
                  status: "cancelled"
                  cancellation_reason: "customer_request"
                  notes: "Customer requested cancellation before shipping"

      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
                  changes:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderChange'
                  metadata:
                    type: object
                    properties:
                      updated_by:
                        type: string
                        example: "admin@bookverse.com"
                      update_reason:
                        type: string
                        example: "Status update - order shipped"
                      processing_time_ms:
                        type: integer
                        example: 89
              examples:
                status_updated:
                  summary: Order status successfully updated
                  value:
                    data:
                      id: "order_789abc123def"
                      status: "shipped"
                      tracking_number: "1Z999AA1234567890"
                      updated_at: "2024-01-26T14:30:00Z"
                    changes:
                      - field: "status"
                        old_value: "processing"
                        new_value: "shipped"
                        timestamp: "2024-01-26T14:30:00Z"
                      - field: "tracking_number"
                        old_value: null
                        new_value: "1Z999AA1234567890"
                        timestamp: "2024-01-26T14:30:00Z"
                    metadata:
                      updated_by: "fulfillment@bookverse.com"
                      update_reason: "Order shipped via UPS Ground"
                      processing_time_ms: 89

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Invalid status transition or business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: "https://docs.bookverse.com/errors/invalid-status-transition"
                title: "Invalid Status Transition"
                status: 409
                detail: "Cannot change status from 'delivered' to 'processing'"
                instance: "/orders/order_789abc123def"
                current_status: "delivered"
                requested_status: "processing"
                valid_transitions: ["returned"]
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Order:
      type: object
      required:
        - id
        - customer_id
        - status
        - payment_status
        - total_amount
        - currency
        - created_at
        - updated_at
      properties:
        id:
          type: string
          pattern: '^order_[a-zA-Z0-9]{12}$'
          description: Unique order identifier
          example: "order_789abc123def"
          readOnly: true
        customer_id:
          type: string
          maxLength: 100
          description: Customer identifier
          example: "customer_12345"
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, returned]
          description: Current order status
          example: "confirmed"
        payment_status:
          type: string
          enum: [pending, authorized, captured, failed, refunded, partially_refunded]
          description: Payment processing status
          example: "captured"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          description: Order items
        subtotal:
          type: number
          format: decimal
          minimum: 0
          description: Subtotal before taxes and shipping
          example: 126.97
        tax_amount:
          type: number
          format: decimal
          minimum: 0
          description: Tax amount
          example: 11.43
        shipping_amount:
          type: number
          format: decimal
          minimum: 0
          description: Shipping cost
          example: 9.99
        discount_amount:
          type: number
          format: decimal
          minimum: 0
          description: Total discount amount
          example: 12.70
        total_amount:
          type: number
          format: decimal
          minimum: 0
          description: Final total amount
          example: 135.69
        currency:
          type: string
          enum: [USD, EUR, GBP, CAD]
          description: Order currency
          example: "USD"
        shipping_address:
          $ref: '#/components/schemas/Address'
        billing_address:
          $ref: '#/components/schemas/Address'
        notes:
          type: string
          maxLength: 500
          description: Customer notes
          example: "Please handle with care"
        internal_notes:
          type: string
          maxLength: 1000
          description: Internal processing notes
          example: "Customer service escalation #12345"
        tracking_number:
          type: string
          maxLength: 50
          description: Shipping tracking number
          example: "1Z999AA1234567890"
        estimated_delivery:
          type: string
          format: date-time
          description: Estimated delivery date
          example: "2024-02-01T00:00:00Z"
        created_at:
          type: string
          format: date-time
          description: Order creation timestamp
          example: "2024-01-25T16:30:00Z"
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-26T09:15:00Z"
          readOnly: true
...
    PaginationInfo:
      type: object
      properties:
        limit:
          type: integer
          description: Results per page
          example: 20
        offset:
          type: integer
          description: Number of results skipped
          example: 0
        total:
          type: integer
          description: Total number of results
          example: 2847
        has_more:
          type: boolean
          description: Whether more results are available
          example: true
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://docs.bookverse.com/errors/validation-error"
            title: "Validation Error"
            status: 400
            detail: "Invalid order data provided"
            instance: "/orders"
...